+++ docs/conf.py
@@ -20,15 +20,15 @@
 # -- Project information -----------------------------------------------------
 
 project = 'StationXML'
-copyright = '2020, International FDSN'
+copyright = '2022, International FDSN'
 author = 'International FDSN'
 
 # The full version, including alpha/beta/rc tags
-version = '1.1.0'
+version = '1.2'
 
 # Documentation version, schema + date
-# ALSO UPDATE the release documentation version in introduction.rst
-doc_version = version + ' (2020-09-02)'
+# ALSO UPDATE the release documentation version in overview.rst
+doc_version = version + ' (2022-02-25)'
 
 # Allow |doc_version| to be used in RST
 rst_epilog = '.. |doc_version| replace:: %s' % doc_version
@@ -40,7 +40,12 @@ rst_epilog = '.. |doc_version| replace:: %s' % doc_version
 # ones.
 extensions = ['sphinxcontrib.contentui',
               'linuxdoc.rstFlatTable',      # Implementation of the 'flat-table' reST-directive
+              #'sphinxcontrib.spelling',
              ]
+
+spelling_word_list_filename=['spelling/text_words.txt',
+                             'spelling/schema_words.txt',
+                             ]
 
 latex_maketitle = r'''
   \noindent\rule{\textwidth}{1pt}\par
@@ -84,7 +89,8 @@ templates_path = ['_templates']
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store','example_1.rst','examples',
+# Note, some of these, like level-*, are included from other .rst files
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store','examples',
                     'level-preamble.rst','level-network.rst','level-channel.rst',
                     'level-response.rst','level-station.rst','response-practical.rst']
 
@@ -120,4 +126,3 @@ html_css_files = [
 html_js_files = [
   'js/sidebar_context.js'
 ]
-