+++ docs/convert_xsd_to_rst.py
@@ -23,8 +23,9 @@ from xmlschema.validators.attributes import XsdAttributeGroup, XsdAnyAttribute
 from xmlschema.validators.complex_types import XsdComplexType
 from xmlschema.validators.facets import *
 
+from xml.etree import ElementTree
 
-def write_tree(element, outfile, first_time = True):
+def write_tree(element, stop_element, outfile, first_time = True):
     """Convert to RST and write element and children to output file.
 
     Starting at `element` recursively operate on all child elements,
@@ -58,90 +59,104 @@ def write_tree(element, outfile, first_time = True):
     print(".. _%s:\n" % href, file=outfile)
 
     # MTH: Hack to fix: SampleRate is not required *unless* SampleRateRatio is present:
+    # Polynomial not required in Stage as in choice
     elements_in_groups = ['SampleRate', 'SampleRateRatio', 'FrequencyStart', 'FrequencyEnd',
-                          'FrequencyDBVariation']
+                          'FrequencyDBVariation', 'Polynomial']
     if element.name in elements_in_groups:
         element.required = False
 
-    if element.required:
+    if element.isRequired():
         print("<%s>     :red:`required`" % (element.name), file=outfile)
     else:
         print("<%s>" % element.name, file=outfile)
 
+    if element.level > len(headings):
+        print(f"level > headings: {element.level} {len(headings)}  {element.name}  {href}")
     print(headings[element.level - 1]*60, file=outfile)
 
     print(".. container:: hatnote hatnote-gray\n", file=outfile)
 
     crumb = element.crumb[0]
+    simplecrumb = f" <{crumb}>"
     for c in element.crumb[1:]:
         crumb = crumb + " :raw-html:`&rarr;`:raw-latex:`$\\rightarrow$` " + c
+        simplecrumb += f" <{c}>"
 
     if not first_time:
         print("   .. container:: crumb\n", file=outfile)
         print("      %s\n" % crumb, file=outfile)
         #print("      crumb:%s\n" % crumb, file=outfile)
 
-    examples = []
-    warnings = []
-    descriptions = []
-
+    description = ""
     stored_description_default=False
-    stored_example_default=False
+    level_char = element.crumb[0][0].upper()
 
     if element.annotation:
-        for note in element.annotation:
-            old_note=note[:]
-            #note=choiceChooser(note,element)
-            note=isRightChoice(note,element)
-            if note:
-                if note==old_note:
-                    isDefault=True
-                else:
-                    isDefault=False
-
-                example_location=note.find("Example:")
-                if example_location != -1:
-                    if not isDefault:
-                         if stored_example_default:
-                             examples.clear()
-                         examples.append(note[example_location+8:])
-
-
-                    elif isDefault and len(examples)==0:
-                        examples.append(note[example_location+8:])
-                        stored_example_default=True
-
-                if example_location!=0:
-                    #If we have one that isn't default, and there are ones that are default, replace.
-                    #If we have one that isn't default, and there aren't, just add.
-                    if example_location!=-1:
-                        toAdd=note[:example_location]
+        ann_has_level_choice = list(filter(lambda a: a.tag == "levelDesc" and a.get("LevelChoice") == level_char , element.annotation))
+        if len(ann_has_level_choice) != 0:
+            ann_list = ann_has_level_choice
+        else:
+            ann_list = list(filter(lambda a: a.tag != "levelDesc" , element.annotation))
+        ann_list_lines = []
+        for ann in ann_list:
+            lines = ann.text.strip().replace('\t', '  ').split("\n")
+            if len(lines) > 1:
+                # second non-empty line often has space indent, first can be on element line
+                leading_whitespace=0
+                for l in lines[1:]:
+                    if len(l.strip()) > 0:
+                        leading_whitespace = len(l)-len(l.lstrip(' '))
+                        break
+                white = " "*leading_whitespace
+                lines[0] = white+lines[0].strip()
+                for l in lines:
+                    if len(l.strip()) != 0:
+                        if l[:leading_whitespace] != white:
+                            raise ValueError(f"removing more spaces than there are! {leading_whitespace} from {l}")
+                        l = l[leading_whitespace:]
                     else:
-                        toAdd=note
+                        l = ""
+                    ann_list_lines.append(f"      {l}\n")
+            else:
+                ann_list_lines.append(f"      {lines[0]}\n")
+            # blank line between annotations
+            ann_list_lines.append("\n")
+#        description = " ".join(map(lambda note: " ".join(note.text.split()) , ann_list))
+        description = "".join(ann_list_lines)
 
-                    if not isDefault:
-                         if stored_description_default:
-                             descriptions.clear()
-                         descriptions.append(toAdd)
+    num_el_attr_warnings = len(element.warning)
+    for attrib in element.attributes:
+        num_el_attr_warnings += len(attrib.warning)
+    if num_el_attr_warnings > 0:
+        with open("warnings.rst", "a") as warnfile:
+            for warning in element.warning:
+                single_line_warning = " ".join(warning.text.strip().split())
+                print("   .. admonition:: Warning, Future Change\n", file=outfile)
+                print(f"      <{element.name}>: {single_line_warning}\n", file=outfile)
 
-                    elif isDefault and len(descriptions)==0:
-                        descriptions.append(toAdd)
-                        stored_description_default=True
+                print(f"\n\n", file=warnfile)
+                print(f"  -    {simplecrumb} : \n", file=warnfile)
+                print("     .. admonition:: Warning, Future Change\n", file=warnfile)
+                print(f"       <{element.name}>: {single_line_warning}\n", file=warnfile)
+            for attrib in element.attributes:
+                for warning in attrib.warning:
+                    single_line_warning = " ".join(warning.text.strip().split())
+                    print( "   .. admonition:: Warning, Future Change\n", file=outfile)
+                    print(f"      {attrib.name}: {single_line_warning}\n", file=outfile)
 
-                if note.find("Warning:") == 0:
-                    warnings.append(note[8:])
+                    print(f"\n\n", file=warnfile)
+                    print(f"  -    {simplecrumb} {attrib.name} : \n", file=warnfile)
+                    print("     .. admonition:: Warning, Future Change\n", file=warnfile)
+                    print(f"       {attrib.name}: {single_line_warning}\n", file=warnfile)
 
-    for warning in warnings:
-        print("   .. admonition:: Warning\n", file=outfile)
-        print("      %s\n" % warning, file=outfile)
 
     if element.type:
         print("   .. container:: type\n", file=outfile)
         latex_words="\t\t\t.. only:: latex\n\n"
         html_words="\t\t\t.. only:: html\n\n"
 
-        latex_words+="\t\t\t\t\ttype: :ref:`%s<type-glossary>`" % (element.type)
-        html_words+="\t\t\t\t\ttype:`%s <appendices.html#glossary-%s>`_" % (element.type,element.type.lower())
+        latex_words+="\t\t\t\t\tcontent type: :ref:`%s<type-glossary>`" % (element.type)
+        html_words+="\t\t\t\t\tcontent type: `%s <appendices.html#glossary-%s>`_" % (element.type,element.type.lower())
         if element.range_string:
             splitElements=element.range_string.split(" ")
 
@@ -153,8 +168,8 @@ def write_tree(element, outfile, first_time = True):
                     result.append(":math:`%s`" % el)
             result=" ".join(result)
 
-            latex_words+=" range:"+result+"\n"
-            html_words+=" range:"+result+"\n"
+            latex_words+="\n\n\t\t\t\t\trange: "+result+"\n"
+            html_words+="\n\n\t\t\t\t\trange: "+result+"\n"
 
         else:
             latex_words+="\n"
@@ -163,38 +178,53 @@ def write_tree(element, outfile, first_time = True):
         print(latex_words, file=outfile)
         print(html_words, file=outfile)
 
-    if len(descriptions) > 0:
-
-        description=descriptions[0]
+    if len(description) > 0:
 
         # If the annotation involves multiple options depending on its location
 
-        #description=choiceChooser(description,element)
         description=urlInserter(description)
         description=mathBlock(description,element.level)
 
-        # Remove whitespace from beginning and end
-        description=description.strip()
-
-        # Add period if needs it
         print("   .. container:: description\n", file=outfile)
-        if description[-1]==".":
-            print("      %s\n" % description, file=outfile)
-        else:
-            print("      %s.\n" % description, file=outfile)
+        print("%s\n" % description, file=outfile)
 
-    # Add period if needs it
-    for example in examples:
-        #example=choiceChooser(example,element)
-        print("   .. container:: example\n", file=outfile)
-        if (example[-1]==">" or example[-1]=="."):
-            print("      **Example**: %s\n" % example, file=outfile)
+    for example in element.example:
+        exampleStr = ""
+        if isinstance(example, ElementTree.Element):
+            if example.get('ElementChoice') is not None and element.parent is not None:
+                if example.get('ElementChoice') != element.parent.name:
+                    # skip this example as wrong level
+                    continue
+            if example.get('LevelChoice') is not None and element.parent is not None:
+                if example.get('LevelChoice') != level_char:
+                    # skip this example as wrong level
+                    continue
+            if example.tag == "example":
+                for ee in example:
+                    exampleStr += ElementTree.tostring(ee, encoding='unicode', method='xml')
+            else:
+                exampleStr = ElementTree.tostring(example, encoding='unicode', method='xml')
         else:
-            print("     **Example**: %s.\n" % example, file=outfile)
+            exampleStr = example
+        exampleStr = exampleStr.strip()
+        print("   .. container:: example\n", file=outfile)
+        if exampleStr.find('\n') != -1:
+            # multiline example
+            print("      **Example**::\n", file=outfile)
+            lines = exampleStr.splitlines()
+            for l in lines:
+                print(f"        {l.rstrip()}", file=outfile)
+        else:
+            if (exampleStr[-1]==">" or exampleStr[-1]=="."):
+                print("      **Example**: %s\n" % exampleStr, file=outfile)
+            else:
+                print("     **Example**: %s.\n" % exampleStr, file=outfile)
 
     if element.attributes:
-        print(".. tabularcolumns::|l|l|l|1|1| \n", file=outfile)
-        print(".. csv-table::", file=outfile)
+        print("\n\n", file=outfile)
+        print(f"   **Attributes of <{element.name}>**: \n", file=outfile)
+        print("   .. tabularcolumns::|l|l|l|1|1| \n", file=outfile)
+        print("   .. csv-table::", file=outfile)
         print("      :class: rows", file=outfile)
         print("      :escape: \ ", file=outfile)
         print('      :header: "attribute", "type", "required", "description", "example"', file=outfile)
@@ -207,69 +237,76 @@ def write_tree(element, outfile, first_time = True):
 
             description = ""
             example = ""
-
-            for note in attrib.annotation:
-                old_note=note[:]
-                #note=choiceChooser(note,element)
-                note=isRightChoice(note,element)
-                if note:
-
-                    if note==old_note:
-                        isDefault=True
+            if len(attrib.example) != 0:
+                # use first example unless ElementChoice is attribute of example
+                DQ='"'
+                example = f"{attrib.name}=\{DQ}{attrib.example[0].text}\{DQ}"
+                for ex in attrib.example:
+                    if ex.get("ElementChoice") is not None and ex.get("ElementChoice") == element.name:
+                        example = f"{attrib.name}=\{DQ}{ex.text}\{DQ}"
+                    if ex.get("LevelChoice") is not None and ex.get("LevelChoice") == level_char:
+                        example = f"{attrib.name}=\{DQ}{ex.text}\{DQ}"
+            if len(attrib.annotation) != 0:
+                    ann_has_level_choice = list(filter(lambda a: a.tag == "levelDesc" and a.get("LevelChoice") == level_char , attrib.annotation))
+                    if len(ann_has_level_choice) != 0:
+                        ann_list = ann_has_level_choice
                     else:
-                        #print(note+" "+old_note, file=sys.stderr)
-                        isDefault=False
+                        ann_list = list(filter(lambda a: a.tag != "levelDesc" , attrib.annotation))
+                    description = " ".join(map(lambda note: " ".join(note.text.split()) , ann_list))
 
-                    example_location=note.find("Example:")
-                    if example_location != -1:
-                        if not isDefault:
-                             example = note[example_location+8:]
-
-                        elif isDefault and len(example)==0:
-                            example = note[example_location+8:]
-
-                    if example_location != 0:
-                        if example_location!=-1:
-                            toAdd= note[:example_location]
-                        else:
-                            toAdd= note
-
-                        if not isDefault:
-                             description = toAdd
-
-                        elif isDefault and len(description)==0:
-                            description = toAdd
-
+            description = " ".join(description.split())
             print("      **%s**, :ref:`%s<type-glossary>`, %s, \"%s\", \"%s\" " % (attrib.name, attrib.type,required, description, example), file=outfile)
 
         print(file=outfile)
 
     if element.children:
+        print("\n\n", file=outfile)
+        print(f"   **Sub Elements of <{element.name}>**: \n", file=outfile)
+        print("   .. tabularcolumns::|l|l|l|l| \n", file=outfile)
+        print("   .. csv-table::", file=outfile)
+        print("      :class: rows", file=outfile)
+        print("      :escape: \ ", file=outfile)
+        print('      :header: "element", "type", "number"', file=outfile)
+        print("      :widths: auto\n", file=outfile)
         for child in element.children:
-            write_tree(child, outfile, first_time = False)
+            parentRef = ""
+            if child.name != stop_element:
+                for c in element.crumb:
+                    parentRef += c+"-"
+            required = ""
+            if child.isRequired():
+                required = ":red:`required`"
+                if child.max_occurs == "unbounded" or child.max_occurs is None:
+                    required = ":red:`required, many`"
+            else:
+                if child.min_occurs == 0:
+                    if child.max_occurs == "unbounded" or child.max_occurs is None:
+                        required = "optional, many"
+                    else:
+                        required = "optional"
+                elif child.max_occurs == 1:
+                    required = "optional"
+                elif child.max_occurs == "unbounded" or child.max_occurs is None:
+                    required = "many"
+                else:
+                    required = f"{child.min_occurs}/{child.max_occurs}"
+
+            range = ""
+            if child.range_string is not None:
+                range = child.range_string
+            type = ""
+            if child.type is not None:
+                type = child.type
+
+            #print(f"      **{child.local_name}**, :ref:`{child.local_name}`, {child.type}, \"{child.required}\", \"{child.range_string}\" ", file=outfile)
+            print(f"      :ref:`\<{child.name}\><{parentRef}{child.name}>`, {type}, \"{required}\" ", file=outfile)
+        print("\n\n", file=outfile)
+
+        for child in element.children:
+            if child.name != stop_element:
+                write_tree(child, stop_element, outfile, first_time = False)
 
     return
-
-
-# when given a documentation (Typically Example or Description) and the element, and it starts with default, will
-# return the correct documentation based off where it is in the xsd
-# Default:Default thing
-# Choice:N Documentation for network
-# Choice:S Documentation for station
-def choiceChooser(documentation,theElement):
-    if documentation.find("LevelDefault:")==0:
-        documentationChoices=documentation.split("LevelChoice:")
-        documentation=documentationChoices[0][13:]
-        sectionSymbol=theElement.crumb[0][0].upper()
-
-        for choice in documentationChoices[1:]:
-            if sectionSymbol==choice[0]:
-                documentation=choice[2:]
-    elif documentation.find("ElementDefault:")==0:
-        documentation=elementChoiceChooser(documentation,theElement)
-
-    return documentation
-
 
 def elementChoiceChooser(documentation,theElement):
     documentationChoices=documentation.split("ElementChoice:")
@@ -369,6 +406,8 @@ class Attribute(object):
         self.local_name = local_name
         self.required = required
         self.annotation = []
+        self.example = []
+        self.warning = []
         self.type = type
 
         if annotation is not None:
@@ -377,6 +416,12 @@ class Attribute(object):
 
     def add_annotation(self, note):
         self.annotation.append(note)
+
+    def add_example(self, ex):
+        self.example.append(ex)
+
+    def add_warning(self, element):
+        self.warning.append(element)
 
     def __repr__(self):
 
@@ -402,12 +447,17 @@ class Element(object):
         self.local_name = local_name
         self.required = required
         self.annotation = []
+        self.example = []
+        self.warning = []
         self.children = []
         self.attributes = []
         self.crumb = []
         self.level = level
         self.type = type
         self.range_string = range_string
+        self.parent = None
+        self.min_occurs = 1
+        self.max_occurs = 1
 
         if annotation is not None:
             for note in annotation:
@@ -424,14 +474,30 @@ class Element(object):
     def add_annotation(self, note):
         self.annotation.append(note)
 
+    def add_example(self, element):
+        self.example.append(element)
+
+    def add_warning(self, element):
+        self.warning.append(element)
+
     def add_attribute(self, attribute):
         self.attributes.append(attribute)
 
     def add_child(self, element):
         self.children.append(element)
+        element.parent = self
 
     def add_crumb(self, x):
         self.crumb.append(x)
+    def isRequired(self):
+        # MTH: Hack to fix: SampleRate is not required *unless* SampleRateRatio is present:
+        elements_in_groups = ['SampleRate', 'SampleRateRatio', 'FrequencyStart', 'FrequencyEnd',
+                              'FrequencyDBVariation']
+        if self.name in elements_in_groups:
+            return False
+        else:
+            return self.required
+
 
 
 def get_type(x, field_name):
@@ -479,8 +545,8 @@ def get_type(x, field_name):
     return keep_type, range_string
 
 
-def walk_tree(xsd_element, stop_element, level=1, last_elem=None, context=None):
-    """Recursively walk the schema element tree until stop_element is encountered"""
+def walk_tree(xsd_element, level=1, last_elem=None, context=None):
+    """Recursively walk the schema element tree """
 
     level_elem = None
     space = (level - 1) * 8 + 6
@@ -491,17 +557,15 @@ def walk_tree(xsd_element, stop_element, level=1, last_elem=None, context=None):
         context = []
 
     if args.verbose > 1:
-        print (f'Working on xsd_element: {xsd_element}, level: {level}, stop_element: {stop_element}')
-
-    # Recursion is complete if the stop element is reached
-    if stop_element and xsd_element.local_name == stop_element:
-        return level_elem
+        print (f'Working on xsd_element: {xsd_element}, level: {level}')
 
     # Store the context for breadcrumbs
     context.append(xsd_element.local_name)
 
-    keep_type, range_string = get_type(xsd_element.type.content_type, xsd_element.local_name)
+    keep_type, range_string = get_type(xsd_element.type.content, xsd_element.local_name)
     this_elem = Element(name=xsd_element.local_name, level=level, type=keep_type, range_string=range_string, crumb=context)
+    this_elem.min_occurs = xsd_element.min_occurs
+    this_elem.max_occurs = xsd_element.max_occurs
 
     if xsd_element.min_occurs == 1:
         this_elem.required = True
@@ -515,15 +579,27 @@ def walk_tree(xsd_element, stop_element, level=1, last_elem=None, context=None):
 
     if xsd_element.annotation is not None:
         for y in xsd_element.annotation.documentation:
-            text = " ".join(y.text.split())
-            this_elem.add_annotation(text)
+            for ex in y.findall("example"):
+                this_elem.add_example(ex)
+            for ex in y.findall("warning"):
+                this_elem.add_warning(ex)
+            for ld in y.findall("levelDesc"):
+                this_elem.add_annotation(ld)
+            if len(y) == 0 and y.text != None:
+                this_elem.add_annotation(y)
 
     # May not want these:
     if xsd_element.type.annotation is not None:
         for y in xsd_element.type.annotation.documentation:
-            text = " ".join(y.text.split())
-            this_elem.add_annotation(text)
-            #print("%s MTH: Add (type) annotation:[%s]" % (space, text), file=sys.stderr)
+            for ex in y.findall("example"):
+                this_elem.add_example(ex)
+            for ex in y.findall("warning"):
+                this_elem.add_warning(ex)
+            for ld in y.findall("levelDesc"):
+                this_elem.add_annotation(ld)
+            if len(y) == 0 and y.text is not None:
+                this_elem.add_annotation(y)
+                #print("%s MTH: Add (type) annotation:[%s]" % (space, text), file=sys.stderr)
 
     for k in xsd_element.attributes:
         attrib = xsd_element.attributes[k]
@@ -534,16 +610,22 @@ def walk_tree(xsd_element, stop_element, level=1, last_elem=None, context=None):
 
         if attrib.annotation is not None:
             for y in attrib.annotation.documentation:
-                text = " ".join(y.text.split())
-                attr.add_annotation(text)
-                #print("%s MTH: Add attrib annotation:[%s]" % (space, text), file=sys.stderr)
+                for ex in y.findall("example"):
+                    attr.add_example(ex)
+                for ex in y.findall("warning"):
+                    attr.add_warning(ex)
+                for ld in y.findall("levelDesc"):
+                    attr.add_annotation(ld)
+                if len(y) == 0 and y.text is not None:
+                    attr.add_annotation(y)
+                    #print("%s MTH: Add attrib annotation:[%s]" % (space, text), file=sys.stderr)
 
     # xsd_element.type is always some form of XsdComplexType
     # xsd_element.type.content_type is either XsdGroup, XsdAtomicBuiltin or XsdAtomicRestriction
-    if isinstance(xsd_element.type.content_type, XsdGroup):
+    if isinstance(xsd_element.type.content, XsdGroup):
 
         # Iterate through contained elements
-        for e in xsd_element.type.content_type.iter_elements():
+        for e in xsd_element.type.content.iter_elements():
 
             if not isinstance(e, XsdAnyElement):
 
@@ -558,14 +640,21 @@ def walk_tree(xsd_element, stop_element, level=1, last_elem=None, context=None):
                     # Store subelement, while adding name to context
                     subElement = Element(name=e.local_name, required=required, type=keep_type, level=level+1,
                                          range_string=range_string, crumb=context + [e.local_name])
+                    subElement.min_occurs = e.min_occurs
+                    subElement.max_occurs = e.max_occurs
                     this_elem.add_child(subElement)
 
                     if e.annotation is not None:
                         for y in e.annotation.documentation:
-                            text = " ".join(y.text.split())
-                            subElement.add_annotation(text)
-                            # print("  has annotation=[%s]" % text, file=sys.stderr)
-
+                            for ex in y.findall("example"):
+                                subElement.add_example(ex)
+                            for ex in y.findall("warning"):
+                                subElement.add_warning(ex)
+                            for ld in y.findall("levelDesc"):
+                                subElement.add_annotation(ld)
+                            if len(y) == 0 and y.text != None:
+                                subElement.add_annotation(y)
+                                # print("  has annotation=[%s]" % text, file=sys.stderr)
                     for k in e.attributes:
                         attrib = e.attributes[k]
                         if not isinstance(attrib, XsdAnyAttribute):
@@ -575,11 +664,17 @@ def walk_tree(xsd_element, stop_element, level=1, last_elem=None, context=None):
 
                         if attrib.annotation is not None:
                             for y in attrib.annotation.documentation:
-                                text = " ".join(y.text.split())
-                                attr.add_annotation(text)
+                                for ex in y.findall("example"):
+                                    attr.add_example(ex)
+                                for ex in y.findall("warning"):
+                                    attr.add_warning(ex)
+                                for ld in y.findall("levelDesc"):
+                                    subElement.add_annotation(ld)
+                                if len(y) == 0 and y.text != None:
+                                    attr.add_annotation(y)
 
                 else:
-                    walk_tree(e, stop_element, level=level+1, last_elem=this_elem, context=context)
+                    walk_tree(e, level=level+1, last_elem=this_elem, context=context)
                     context.pop()
 
     else:
@@ -587,6 +682,28 @@ def walk_tree(xsd_element, stop_element, level=1, last_elem=None, context=None):
         return level_elem
 
     return level_elem
+
+def save_spelling(words):
+    spelling_dir = 'spelling'
+    schema_words = 'schema_words.txt'
+    text_words = 'text_words.txt'
+    if not os.path.isdir(spelling_dir):
+        os.makedirs(spelling_dir)
+    with open(os.path.join(spelling_dir, schema_words), 'w') as words_file:
+        sort_words = list(words)
+        sort_words.sort()
+        for w in sort_words:
+            print(w, file=words_file)
+
+
+def recur_spelling(words, element):
+    words.add(element.name)
+    for a in element.attributes:
+        words.add(a.name)
+    for child in element.children:
+        recur_spelling(words, child)
+
+
 
 
 def main():
@@ -604,12 +721,22 @@ def main():
 
     schema = xmlschema.XMLSchema(args.schemafile)
 
+    # create warning.rst file for deprectaion warnings
+    with open("warnings.rst", "w") as warnfile:
+        print(".. Put any comments here\n", file=warnfile)
+        print("  Warning, this file is regenerated from the annotations in the schema file.\n", file=warnfile)
+        print("  Any changes will be overwritten by convert_xsd_to_rst.py.\n\n\n", file=warnfile)
+        print("\n", file=warnfile)
+        print("The following are potential future changes, as tagged in the schema with <warning> elements in the documentation. They may result in modifications or deletions in future versions of StationXML.\n", file=warnfile)
+        print("\n\n", file=warnfile)
+
     level_xpaths = ['fsx:FDSNStationXML',
                     'fsx:FDSNStationXML/fsx:Network',
                     'fsx:FDSNStationXML/fsx:Network/fsx:Station',
                     'fsx:FDSNStationXML/fsx:Network/fsx:Station/fsx:Channel',
                     'fsx:FDSNStationXML/fsx:Network/fsx:Station/fsx:Channel/fsx:Response']
 
+    words = set()
     for i, xpath in enumerate(level_xpaths):
         xsd_element = schema.find(xpath)
 
@@ -618,7 +745,7 @@ def main():
         if i < 4:
             stop_element = os.path.basename(level_xpaths[i+1]).split(':')[1]
 
-        level_elem = walk_tree(xsd_element, stop_element)
+        level_elem = walk_tree(xsd_element)
 
         # Use Preamble instead of root element name
         if this_element=="FDSNStationXML":
@@ -631,7 +758,14 @@ def main():
             print (f'Writing to {rst_file}')
 
         with open(rst_file, 'w') as outfile:
-            write_tree(level_elem, outfile = outfile)
+            print(".. Put any comments here\n", file=outfile)
+            print("  Warning, this file is regenerated from the annotations in the schema file.\n", file=outfile)
+            print("  Any changes will be overwritten by convert_xsd_to_rst.py.\n", file=outfile)
+
+            write_tree(level_elem, stop_element, outfile = outfile)
+
+        recur_spelling(words, level_elem)
+    save_spelling(words)
 
 
 if __name__ == "__main__":